Extra and alias Cursor, reorg package definitions
