+++ indigo/indigo-extras/src/main/scala/indigoextras/ui/shaders/package.scala
@@ -1,0 +1,11 @@
+package indigoextras.ui.shaders
+
+import indigo.shared.shader.ShaderProgram
+
+val ui: Set[ShaderProgram] =
+  Set(
+    indigoextras.ui.shaders.LayerMask.shader
+  )
+
+val all: Set[ShaderProgram] =
+  ui
